Work on #3102. Moved all the components out of model.jsx into their own files so they can be grabbed via getComponent()
