+++ src/core/components/models.jsx
@@ -1,6 +1,5 @@
 import React, { Component, PropTypes } from "react"
 
-
 export default class Models extends Component {
   static propTypes = {
     getComponent: PropTypes.func,
@@ -16,7 +15,7 @@ export default class Models extends Component {
     let { docExpansion } = getConfigs()
     let showModels = layoutSelectors.isShown("models", docExpansion === "full" || docExpansion === "list" )
 
-    const Model = getComponent("model")
+    const ModelWrapper = getComponent("ModelWrapper")
     const Collapse = getComponent("Collapse")
 
     if (!definitions.size) return null
@@ -31,8 +30,9 @@ export default class Models extends Component {
       <Collapse isOpened={showModels} animated>
         {
           definitions.entrySeq().map( ( [ name, model ])=>{
+            console.log("model", name, model)
             return <div className="model-container" key={ `models-section-${name}` }>
-              <Model name={ name }
+              <ModelWrapper name={ name }
                      schema={ model }
                      isRef={ true }
                      getComponent={ getComponent }
